Backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
langchain
langchain-google-genai
shapely
pydantic
websockets
python-dotenv
